langchain/pyproject.toml
[tool.poetry]
name = "langserver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.3.1"}
anyio = "^4.8.0"
langchain-cli = "^0.0.36"
langchain-deepseek = "^0.1.2"
langchain = "^0.3.20"
langchain-core = "^0.3.45"
openai = "1.66.2"
numpy = ">=1.22.4,<2.0.0"
langchain-community = {version = "^0.3.20", source = "tuna"}
langgraph = "^0.3.18"
beautifulsoup4 = "^4.13.3"
chroma-hnswlib = "0.7.6"
langchain-chroma = "^0.2.2"
transformers = "^4.50.0"
sentence-transformers = "^3.4.1"
langchain-huggingface = "^0.1.2"
flagembedding = "^1.3.4"
mysql-connector-python = "^9.2.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
